backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
sqlparse==0.5.3
firebase-admin==6.6.0
django-cors-headers==4.6.0
djangorestframework==3.15.2